Rewrote tests for ParticipantTableGrid.
